use Value for Randomin Eff

+++ src/Interpreter.hs
@@ -136,8 +136,8 @@ run seed story e0 = loop (initState seed pc0) e0 k0
       Random range -> do
         let State{seed} = s
         let x = stepRandom seed
-        let result = x `mod` (range+1) + 1
-        k s { seed = x } result
+        let result = x `mod` (fromIntegral range + 1) + 1
+        k s { seed = x } (fromIntegral result)
 
       Quit -> do
         -- dont call "k" but instead "k0"
